Major refactoring and reengineering of the whole library to add support to features like Log Topics and log _filter_ by log level before writing.
The class now manages the LogSeverity in a SysLog fashion, not only t facilitate the SysLog support in the future but because the author likes this levels more than other log libraries around.

A lot of breaking changes, including that now you have to instantiate a Log to use it (previous versions were based on a record, not a class, that you don't need to instantiate (much like the IOUtils records), but a long standing requirement is to allow a single program to produce different log files and the only way to do it is allowing multiple log objects to be alive at the same time.

The ujachLogAuto.pas unit is introduced to have an auto-setup that is valid for 95% of the projects, so the product keeps an easy way to integrate in any existing Delphi project and required changes to existing code are kept at minimum.

Coming demos will show how to take advantage of this features.