In periods of high log load, the routine that updates the RichEdit control hangs up, as sometimes log entries are added at a fast peace to the queue than to the control and the routine never ends.
It also re-enters immediately if the timer expired during the update period.

For that, the timer disabled at entry and re-enabled at the end of the timer event, so it gives enough time to other activities the main thread must do in order to be responsive and the new MaxMillisecondsAddingMessages property is used to control how much time is allowed this routine to add messages (it may take a bit more, since after reaching the limit it performs some additional tasks).

The new routine was tested on high load and now works much better than before, in combination with auto-cap the entire test program responds very well to the user input and control.

Resolves #27