Factorized the iteration trough the entry queue to allow the writers to override it, for example to allow a bulk insert to a database
This is a must to achieve good performance while writing logs to a database connection.

With the current OpenLogChannel/CloseLogChannel approach, the writer would have to cache the log entries data passed as parameters to the write method. It is better to give the writer access to the queue and allow it to work with or cache the entries directly.

Resolves #40